common: separating epoll to its own header file
next patch(es) will change epoll implementation to use slist. This
separation of the  header file is necessary to include ofi_signal.h and
ofi_list.h without circular dependency.
