+++ include/ofi_signal.h
@@ -113,91 +113,5 @@ static inline int fd_signal_poll(struct fd_signal *signal, int timeout)
 	return (ret == 0) ? -FI_ETIMEDOUT : 0;
 }
 
-#ifdef HAVE_EPOLL
-#include <sys/epoll.h>
-
-#define FI_EPOLL_IN  EPOLLIN
-#define FI_EPOLL_OUT EPOLLOUT
-
-typedef int fi_epoll_t;
-
-static inline int fi_epoll_create(int *ep)
-{
-	*ep = epoll_create(4);
-	return *ep < 0 ? -ofi_syserr() : 0;
-}
-
-static inline int fi_epoll_add(int ep, int fd, uint32_t events, void *context)
-{
-	struct epoll_event event;
-	int ret;
-
-	event.data.ptr = context;
-	event.events = events;
-	ret = epoll_ctl(ep, EPOLL_CTL_ADD, fd, &event);
-	if ((ret == -1) && (ofi_syserr() != EEXIST))
-		return -ofi_syserr();
-	return 0;
-}
-
-static inline int fi_epoll_mod(int ep, int fd, uint32_t events, void *context)
-{
-	struct epoll_event event;
-
-	event.data.ptr = context;
-	event.events = events;
-	return epoll_ctl(ep, EPOLL_CTL_MOD, fd, &event) ? -ofi_syserr() : 0;
-}
-
-static inline int fi_epoll_del(int ep, int fd)
-{
-	return epoll_ctl(ep, EPOLL_CTL_DEL, fd, NULL) ? -ofi_syserr() : 0;
-}
-
-static inline int fi_epoll_wait(int ep, void **contexts, int max_contexts,
-                                int timeout)
-{
-	struct epoll_event events[max_contexts];
-	int ret;
-	int i;
-
-	ret = epoll_wait(ep, events, max_contexts, timeout);
-	if (ret == -1)
-		return -ofi_syserr();
-
-	for (i = 0; i < ret; i++)
-		contexts[i] = events[i].data.ptr;
-	return ret;
-}
-
-static inline void fi_epoll_close(int ep)
-{
-	close(ep);
-}
-
-#else
-#include <poll.h>
-
-#define FI_EPOLL_IN  POLLIN
-#define FI_EPOLL_OUT POLLOUT
-
-typedef struct fi_epoll {
-	int		size;
-	int		nfds;
-	struct pollfd	*fds;
-	void		**context;
-	int		index;
-	struct fd_signal signal;
-} *fi_epoll_t;
-
-int fi_epoll_create(struct fi_epoll **ep);
-int fi_epoll_add(struct fi_epoll *ep, int fd, uint32_t events, void *context);
-int fi_epoll_mod(struct fi_epoll *ep, int fd, uint32_t events, void *context);
-int fi_epoll_del(struct fi_epoll *ep, int fd);
-int fi_epoll_wait(struct fi_epoll *ep, void **contexts, int max_contexts,
-                  int timeout);
-void fi_epoll_close(struct fi_epoll *ep);
-
-#endif /* HAVE_EPOLL */
 
 #endif /* _OFI_SIGNAL_H_ */
